+++ changelog.md
@@ -9,6 +9,8 @@
   * Reduced rails crafting
   * Removed Nether Infused Powered Rails
 * Shields plus data pack no longer enables itself if the mod is not present
+* Increased items base damage by 0.5 but reduced enchantments damage
+  * Rage (2 -> 1.5 max damage per level)
 * Moved Season plant growth modifier and livestock to ITR
 * Fixed again time control not working on world creation
 * Minor fixes